# api-docs.yaml
openapi: 3.0.1
info:
  title: Open API documentation for Match Details
  description: Match Details
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/match:
    get:
      tags:
      - match-details-controller
      operationId: getMatchDetail
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MatchDetail'
    post:
      tags:
      - match-details-controller
      operationId: saveMatchDetail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchDetail'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MatchDetail'
  /api/match/{id}:
    patch:
      tags:
      - match-details-controller
      operationId: updateMatchDetails
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonPatch'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/team:
    get:
      tags:
      - team-controller
      operationId: getTeamDetails
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
  /api/match/upcomingmatches:
    get:
      tags:
      - match-details-controller
      operationId: getUpcomingMatchesByTeam
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchDetail'
    delete:
      tags:
      - match-details-controller
      operationId: deleteUpCommingMatches
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/match/team/{id}:
    get:
      tags:
      - match-details-controller
      operationId: getMatchesByTeamId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchDetail'
  /api/match/allwinners:
    get:
      tags:
      - match-details-controller
      operationId: getAllWinnerList
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchDetail'
components:
  schemas:
    MatchDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        matchStartTime:
          type: string
          format: date-time
        homeTeamName:
          type: string
        awayTeamName:
          type: string
        homeTeamScore:
          type: integer
          format: int32
        awayTeamScore:
          type: integer
          format: int32
    JsonPatch:
      type: object
    Team:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string