# src/main/java/com/capitalnumber/demo/model/Team.java
package com.capitalnumber.demo.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class Team {
    private Long id;
    private String name;
}